Created detailed text description for Read.ME
Added title for Readme.MD file. Also added description of website, including that it's a two page blog post website primarily powered by JavaScript, including concepts such as localStorage.setItem(), window.location.href, etc. I also included an image of website, a description of it's functionality, an image of the website's functionality, and finally a link for the deployed application.
"The images and link only have placeholder spots so I'll have to add them in later."